+++ python/packages/foundry/agent_framework_foundry/_to_prompt_agent.py
@@ -12,8 +12,9 @@ definition used for local execution can be published as a hosted prompt agent
 without restating the model deployment name. Generation parameters
 (``temperature``, ``top_p``, ``tool_choice``, ``reasoning``,
 ``response_format`` / ``text`` / ``verbosity``) are translated from
-``agent.default_options`` by ``FoundryChatClient._prepare_prompt_agent_options``
-so they stay consistent with the agent's local request path.
+``agent.default_options`` by the local ``_prepare_prompt_agent_options``
+helper, which reuses the chat client's own request-path helpers so they stay
+consistent with the agent's local execution.
 
 Parameters with no Agent Framework equivalent (``structured_inputs``,
 ``rai_config``) are accepted as keyword arguments only.
@@ -60,10 +61,10 @@ def to_prompt_agent(
     All generation parameters that have an Agent Framework equivalent
     (``temperature``, ``top_p``, ``tool_choice``, ``reasoning``,
     ``response_format`` / ``text`` / ``verbosity``) are sourced from
-    ``agent.default_options`` and translated by
-    ``FoundryChatClient._prepare_prompt_agent_options``. The agent is the
-    single source of truth for these; configure them on the ``Agent`` (or
-    pass ``default_options={...}`` to its constructor) rather than here.
+    ``agent.default_options`` and translated by ``_prepare_prompt_agent_options``.
+    The agent is the single source of truth for these; configure them on the
+    ``Agent`` (or pass ``default_options={...}`` to its constructor) rather
+    than here.
 
     Args:
         agent: An Agent Framework agent whose client is a ``FoundryChatClient``.
@@ -102,7 +103,8 @@ def to_prompt_agent(
         getattr(agent, "mcp_tools", []),
     )
 
-    translated = agent.client._prepare_prompt_agent_options(  # pyright: ignore[reportPrivateUsage]
+    translated = _prepare_prompt_agent_options(
+        agent.client,
         agent.default_options,
         has_tools=bool(tools),
     )
@@ -121,6 +123,115 @@ def to_prompt_agent(
         kwargs["rai_config"] = rai_config
 
     return PromptAgentDefinition(**kwargs)
+
+
+def _prepare_prompt_agent_options(
+    client: RawFoundryChatClient[Any],
+    default_options: Mapping[str, Any],
+    *,
+    has_tools: bool = False,
+) -> dict[str, Any]:
+    """Translate ``default_options`` into ``PromptAgentDefinition`` field kwargs.
+
+    Reuses the chat client's own request-path helpers
+    (``validate_tool_mode``, ``client._prepare_response_and_text_format``,
+    ``type_to_text_format_param``) so a published prompt agent stays
+    consistent with the agent's local execution.
+
+    Only fields with a direct ``PromptAgentDefinition`` counterpart are
+    translated: ``temperature``, ``top_p``, ``reasoning``, ``tool_choice``,
+    ``response_format`` / ``text`` / ``verbosity``. Other ``OpenAIChatOptions``
+    keys (``include``, ``prompt``, ``store``, etc.) have no prompt-agent
+    equivalent and are intentionally ignored. The input mapping is never
+    mutated.
+
+    Args:
+        client: The bound ``FoundryChatClient`` (used to reuse its
+            ``_prepare_response_and_text_format`` for dict-shaped
+            ``response_format`` values).
+        default_options: The agent's ``default_options`` mapping.
+
+    Keyword Args:
+        has_tools: When ``False``, ``tool_choice`` is dropped (no point
+            emitting a tool selection policy when the definition has no
+            tools), mirroring the regular request path in
+            ``_prepare_options``.
+
+    Returns:
+        A dict ready to splat into ``PromptAgentDefinition(**...)``. Unset
+        fields are omitted.
+    """
+    from agent_framework._types import validate_tool_mode
+    from azure.ai.projects.models import (
+        PromptAgentDefinitionTextOptions,
+        Reasoning,
+        ToolChoiceAllowed,
+        ToolChoiceFunction,
+    )
+    from openai.lib._parsing._responses import (  # type: ignore[reportPrivateImportUsage]
+        type_to_text_format_param,
+    )
+    from pydantic import BaseModel
+
+    result: dict[str, Any] = {}
+
+    if (temperature := default_options.get("temperature")) is not None:
+        result["temperature"] = temperature
+    if (top_p := default_options.get("top_p")) is not None:
+        result["top_p"] = top_p
+
+    if (reasoning := default_options.get("reasoning")) is not None:
+        if isinstance(reasoning, Reasoning):
+            result["reasoning"] = reasoning
+        elif isinstance(reasoning, Mapping):
+            result["reasoning"] = Reasoning(**dict(cast("Mapping[str, Any]", reasoning)))
+        else:
+            result["reasoning"] = reasoning
+
+    if has_tools and (tool_choice := default_options.get("tool_choice")) is not None:
+        tool_mode = validate_tool_mode(tool_choice)
+        if tool_mode is not None:
+            mode = tool_mode.get("mode")
+            func_name = tool_mode.get("required_function_name")
+            allowed = tool_mode.get("allowed_tools")
+            if mode == "required" and func_name is not None:
+                result["tool_choice"] = ToolChoiceFunction(name=func_name)
+            elif mode == "auto" and allowed is not None:
+                result["tool_choice"] = ToolChoiceAllowed(
+                    mode="auto",
+                    tools=[{"type": "function", "name": name} for name in allowed],
+                )
+            else:
+                result["tool_choice"] = mode
+
+    existing_text = default_options.get("text")
+    text_config: dict[str, Any] | None = (
+        dict(cast("Mapping[str, Any]", existing_text)) if isinstance(existing_text, Mapping) else None
+    )
+    response_format = default_options.get("response_format")
+    if response_format is not None or text_config is not None:
+        if isinstance(response_format, type) and issubclass(response_format, BaseModel):
+            format_config = dict(type_to_text_format_param(response_format))
+            text_config = dict(text_config) if text_config else {}
+            if "format" in text_config and text_config["format"] != format_config:
+                raise ValueError("Conflicting response_format definitions detected.")
+            text_config["format"] = format_config
+        elif response_format is not None:
+            response_format_model, text_config = client._prepare_response_and_text_format(  # pyright: ignore[reportPrivateUsage]
+                response_format=response_format, text_config=text_config
+            )
+            if response_format_model is not None:
+                raise ValueError(
+                    "response_format must be a Pydantic BaseModel subclass or a mapping when "
+                    "converting to a PromptAgentDefinition."
+                )
+    if (verbosity := default_options.get("verbosity")) is not None:
+        text_config = dict(text_config) if text_config else {}
+        text_config["verbosity"] = verbosity
+    if text_config:
+        result["text"] = PromptAgentDefinitionTextOptions(text_config)
+
+    return result
 
 
 def _convert_tools(
